+++ src/main/java/bridge/domain/BridgeNumberGenerator.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.domain;
 
 @FunctionalInterface
 public interface BridgeNumberGenerator {
